services/cart/requirements.txt
fastapi
uvicorn[standard]
redis[asyncio]
PyJWT
python-dotenv
pydantic
pydantic-settings